Only make the selected element draggy, als add drag overlay
This currently requires an extra click to make it draggy
and then to start moving the element (by its overlay). Later,
we'll have to ditch draggy.js and (re)implement all its
stuff ourselves, to combine our overlay with the dragging.

Alternatively, we need to find a way to redispatch the
click event to the newly added overlay.