per-connection prefix to created elements to reduce the chance of collisions
This doesn't prevent a malicious clients from creating collisions as the
collision ids are somewhat predictable, but at least inadverent
collisions should be impossible.

A potential approach to even better collision avoidance could be to
randomize the prefix on the client for each new element, or have the server
(re)set the prefix on the clients randomly.